Remove github yml, update version, tooltips for A*B+C
