+++ plugin.json
@@ -1,6 +1,6 @@
 {
   "slug": "Befaco",
-  "version": "2.0.0",
+  "version": "2.1.0",
   "license": "GPL-3.0-only",
   "name": "Befaco",
   "brand": "Befaco",
