--- src/noise-plethora/teensy/README
@@ -1,2 +1,0 @@
-
-TODO: restore white noise instance count
